requirements.txt
PyYAML==3.09
lxml==2.3
